# go/pkg/internal/server/server_gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package server

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// CancelRequest defines model for CancelRequest.
type CancelRequest struct {
	TaskId string `json:"task_id"`
}

// ContinueRequest defines model for ContinueRequest.
type ContinueRequest struct {
	Input       interface{} `json:"input"`
	Name        string      `json:"name"`
	ResponseUrl string      `json:"response_url"`
	TaskId      string      `json:"task_id"`
}

// StartRequest defines model for StartRequest.
type StartRequest struct {
	Input       interface{} `json:"input"`
	Name        string      `json:"name"`
	ResponseUrl string      `json:"response_url"`
	TaskId      string      `json:"task_id"`
}

// Task defines model for Task.
type Task struct {
	Name string `json:"name"`
}

// Tasks defines model for Tasks.
type Tasks struct {
	Tasks []Task `json:"tasks"`
}

// PostCancelJSONRequestBody defines body for PostCancel for application/json ContentType.
type PostCancelJSONRequestBody = CancelRequest

// PostContinueJSONRequestBody defines body for PostContinue for application/json ContentType.
type PostContinueJSONRequestBody = ContinueRequest

// PostStartJSONRequestBody defines body for PostStart for application/json ContentType.
type PostStartJSONRequestBody = StartRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Cancel a task
	// (POST /cancel)
	PostCancel(w http.ResponseWriter, r *http.Request)
	// Continue a task
	// (POST /continue)
	PostContinue(w http.ResponseWriter, r *http.Request)
	// Start a task
	// (POST /start)
	PostStart(w http.ResponseWriter, r *http.Request)
	// Get all defined tasks
	// (GET /tasks)
	GetTasks(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Cancel a task
// (POST /cancel)
func (_ Unimplemented) PostCancel(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Continue a task
// (POST /continue)
func (_ Unimplemented) PostContinue(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Start a task
// (POST /start)
func (_ Unimplemented) PostStart(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get all defined tasks
// (GET /tasks)
func (_ Unimplemented) GetTasks(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostCancel operation middleware
func (siw *ServerInterfaceWrapper) PostCancel(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostCancel(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostContinue operation middleware
func (siw *ServerInterfaceWrapper) PostContinue(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostContinue(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostStart operation middleware
func (siw *ServerInterfaceWrapper) PostStart(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostStart(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTasks operation middleware
func (siw *ServerInterfaceWrapper) GetTasks(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTasks(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/cancel", wrapper.PostCancel)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/continue", wrapper.PostContinue)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/start", wrapper.PostStart)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/tasks", wrapper.GetTasks)
	})

	return r
}

type PostCancelRequestObject struct {
	Body *PostCancelJSONRequestBody
}

type PostCancelResponseObject interface {
	VisitPostCancelResponse(w http.ResponseWriter) error
}

type PostCancel200Response struct {
}

func (response PostCancel200Response) VisitPostCancelResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type PostCancel400Response struct {
}

func (response PostCancel400Response) VisitPostCancelResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostContinueRequestObject struct {
	Body *PostContinueJSONRequestBody
}

type PostContinueResponseObject interface {
	VisitPostContinueResponse(w http.ResponseWriter) error
}

type PostContinue200Response struct {
}

func (response PostContinue200Response) VisitPostContinueResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type PostContinue400Response struct {
}

func (response PostContinue400Response) VisitPostContinueResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostContinue401Response struct {
}

func (response PostContinue401Response) VisitPostContinueResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostContinue403Response struct {
}

func (response PostContinue403Response) VisitPostContinueResponse(w http.ResponseWriter) error {
	w.WriteHeader(403)
	return nil
}

type PostContinue404Response struct {
}

func (response PostContinue404Response) VisitPostContinueResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type PostContinue500Response struct {
}

func (response PostContinue500Response) VisitPostContinueResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type PostStartRequestObject struct {
	Body *PostStartJSONRequestBody
}

type PostStartResponseObject interface {
	VisitPostStartResponse(w http.ResponseWriter) error
}

type PostStart202Response struct {
}

func (response PostStart202Response) VisitPostStartResponse(w http.ResponseWriter) error {
	w.WriteHeader(202)
	return nil
}

type PostStart400Response struct {
}

func (response PostStart400Response) VisitPostStartResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostStart401Response struct {
}

func (response PostStart401Response) VisitPostStartResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostStart403Response struct {
}

func (response PostStart403Response) VisitPostStartResponse(w http.ResponseWriter) error {
	w.WriteHeader(403)
	return nil
}

type PostStart500Response struct {
}

func (response PostStart500Response) VisitPostStartResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type GetTasksRequestObject struct {
}

type GetTasksResponseObject interface {
	VisitGetTasksResponse(w http.ResponseWriter) error
}

type GetTasks200JSONResponse Tasks

func (response GetTasks200JSONResponse) VisitGetTasksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetTasks500Response struct {
}

func (response GetTasks500Response) VisitGetTasksResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Cancel a task
	// (POST /cancel)
	PostCancel(ctx context.Context, request PostCancelRequestObject) (PostCancelResponseObject, error)
	// Continue a task
	// (POST /continue)
	PostContinue(ctx context.Context, request PostContinueRequestObject) (PostContinueResponseObject, error)
	// Start a task
	// (POST /start)
	PostStart(ctx context.Context, request PostStartRequestObject) (PostStartResponseObject, error)
	// Get all defined tasks
	// (GET /tasks)
	GetTasks(ctx context.Context, request GetTasksRequestObject) (GetTasksResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// PostCancel operation middleware
func (sh *strictHandler) PostCancel(w http.ResponseWriter, r *http.Request) {
	var request PostCancelRequestObject

	var body PostCancelJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostCancel(ctx, request.(PostCancelRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostCancel")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostCancelResponseObject); ok {
		if err := validResponse.VisitPostCancelResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostContinue operation middleware
func (sh *strictHandler) PostContinue(w http.ResponseWriter, r *http.Request) {
	var request PostContinueRequestObject

	var body PostContinueJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostContinue(ctx, request.(PostContinueRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostContinue")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostContinueResponseObject); ok {
		if err := validResponse.VisitPostContinueResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostStart operation middleware
func (sh *strictHandler) PostStart(w http.ResponseWriter, r *http.Request) {
	var request PostStartRequestObject

	var body PostStartJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostStart(ctx, request.(PostStartRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostStart")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostStartResponseObject); ok {
		if err := validResponse.VisitPostStartResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetTasks operation middleware
func (sh *strictHandler) GetTasks(w http.ResponseWriter, r *http.Request) {
	var request GetTasksRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetTasks(ctx, request.(GetTasksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTasks")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetTasksResponseObject); ok {
		if err := validResponse.VisitGetTasksResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}