# openapi/openapi.yaml
---
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0
paths:
  /start:
    post:
      summary: Start a task
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartRequest'
      responses:
        '202':
          description: Task started
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
  /continue:
    post:
      summary: Continue a task
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinueRequest'
      responses:
        '200':
          description: Task continued
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found
        '500':
          description: Internal server error
  /cancel:
    post:
      summary: Cancel a task
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: Task canceled
        '400':
          description: Bad request
  /tasks:
    get:
      summary: Get all defined tasks
      responses:
        '200':
          description: All defined tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tasks'
        '500':
          description: Internal server error
components:
  schemas:
    StartRequest:
      type: object
      properties:
        name:
          type: string
        input: {}
        response_url:
          type: string
        task_id:
          type: string
      required:
      - name
      - input
      - response_url
      - task_id
    CancelRequest:
      type: object
      properties:
        task_id:
          type: string
      required:
      - task_id
    ContinueRequest:
      type: object
      properties:
        name:
          type: string
        input: {}
        task_id:
          type: string
        response_url:
          type: string
      required:
      - name
      - input
      - task_id
      - response_url
    Tasks:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
      required:
      - tasks
    Task:
      type: object
      properties:
        name:
          type: string
      required:
      - name